Retrieve user-defined regions from Firestore and render on map (#68)
I manually created these regions in FIrestore using the Web UI.

Note that if we deploy tomorrow, these triangles will be rendered on the map, but should be harmless unless the affected region is Houston again, since they won't show up on the map. If it is Houston, we'll have to scrub this test data. Eventually, the user data layer will be toggleable.