Save user regions to backend (#86)
Save user regions when they are first drawn, or when they are deleted, to the database.

Use a different Firestore collection when in a test, so that we can avoid writing to production data. We also delete all test-drawn polygons before and after the test.

Added unit tests, which required mocking Firebase a little bit. Adding the "delete all test-drawn polygons" required adding the Firebase Node package as a dependency.

Moved "score" constant to layer_util.js, so that run.js depends on it, rather than the other way around, since layer_util is a lower-level library.