Switch to using a webserver from yarn (#157)
Google Cloud SDK is a much heavier-weight dependency, and we're not using Google App Engine anyway. This also allows us to reorganize the directory structure so that the webserver's root is the docs/ directory. This is nice because github pages has an option to treat this directory specially.

This is inspired by getting things set up on CI, where a GAE dependency is really a problem, but yarn dependencies are trivial.