Enable Travis CI on github pushes/pull requests (#159)
Main change is .travis.yml addition. yarn.lock is added to the repository to signal to Travis that it should use yarn to manage dependencies rather than npm. Since my .gitignore listed yarn.lock, other people's might too, so please check that and remove yarn.lock from there if necessary, since it's now part of the repository.

Two tests are failing (click_feature_test and polygon_draw_test). I've disabled them on CI and will debug them separately.

Currently, Travis runs on every push and every pull request unless the commit message has '[skip ci]' in it. I think that should be ok, especially since I've set it so that queued requests are canceled if a more recent commit comes in, but we can also only set it for merges to master and pull requests (I think) if we want to.

Progress on #21 (just need to get remaining tests enabled).