Reduce Travis builds to a single machine and wait longer for drawing toolbar (#184)
Pushing triggers enough builds that we often have builds queued. At that point, parallelization isn't very useful because the overhead of starting up an additional machine is very high, and we only have two machines available.

Also wait longer for the "Draw a shape" button to appear, since it is now dependent on Firebase, and the polygon_draw_test doesn't wait for the whole page to load.