Extract script-loading functionality and add EarthEngine support (#183)
Get rid of mock_ee.js in favor of the real EarthEngine and mocking/stubbing the appropriate methods. Not a big change, since we're not actually letting any EarthEngine evaluate() methods run, but I think this is cleaner, and we could add some more integration-y tests that do invoke the genuine EarthEngine evaluate.

Allow tests to specify the external scripts that they need, so that as we add new scripts, tests that don't need them can remain fast. (Might be good to do this for the test logins, but punting for now.)