Restructure Firestore data to avoid document with unknown collections (#190)
As we've discussed offline, the Firebase web client cannot list the collections in a document, which means that, since in our current schema a year is a document, we can't list the disasters from that year. To get around this, make the structure disaster-metadata/<year>-<disaster> and then put all disaster metadata (except for usershapes, which remain global, hmm...) into fields in the <year>-<disaster> document.

I used a script for this that I'd rather not clean up for submission, since this should be a one-off. But if we need to do this again, I can push it up.

Also fix a sequential processing issue: we were blocking score computation, the most expensive part of our page load, on Firestore retrieval, because we wanted to know how many layers there were. But since score always goes last, we can just do a bit of special-casing in layer_util.js to avoid needing to know how many layers there are ahead of time.

I tested that we're actually using the new data by changing a layer to be true by default in the new data and verifying that it was really there.