Move to using a test-only Google account in our tests (#187)
This also required a fair amount of Firebase reworking. I would break it into two or more PRs, but our tests are broken, and this fixes them, so seems worth getting them in as one change.

Changes are:
1. Move to using a test-only Firestore database. This requires:
1. Populating and deleting data from that database for all integration tests (and some unit tests). This leads us to:
1. Do Firestore initialization and cleanup in a plugin, for all integration tests and the unit tests that need it. Doing it in a plugin allows the Firebase Javascript objects to be totally separate, which is nice, and also to do things that you can only do server-side, like find collections recursively so that you can delete them.
1. Also clean up support/index.js, which was getting a bit gnarly, pushing all the mess into another file that we include.

This un-breaks Travis, and exposes a probably already existing failure in click_feature_test that isn't fixed here.