Add approximate SNAP fraction to user-drawn polygons (#192)
Make the expected changes to do EarthEngine computations. We intersect each block group with the user-drawn polygon and see what fraction of the block group is covered by the polygon. Assuming uniform population, we can then calculate the expected population in the polygon intersect that block group, and sum over all block groups to get the expected population of the polygon. The same goes for subpopulations, like SNAP recipients.

Significant changes to unit_tests/polygon_draw_test: we now use real EarthEngine, because these calculations are complex enough that stubbing everything out seems error-prone. Because we then have to wait for EarthEngine callbacks to complete, come good on a plan from a while ago and return Promises from StoredShapeData methods all the way out, so that the test can tell when callbacks are done. Note that we log informative messages when Promises throw errors, so we don't really need to catch/log them the way we often do (unless we want to recover).

Move rendering responsibility of calculatedData from popup.js to polygon_draw.js, since polygon_draw.js is responsible for data calculation, and this keeps it in one place.

Also realized that we don't need a real Firestore database for polygon_draw unit test, just the library, so got rid of that from script_loader.js.