Fix deletion of test data from Firestore (#199)
If the baseUrl is not set, then Cypress runs beforeEach hooks twice (https://github.com/cypress-io/cypress/issues/2636), which left lingering data.

With that fix, the belt-and-suspenders approach is mostly unnecessary, but added it for good measure: make the test directory start with the start date of the test and at the start of each test, delete all directories older than 24 hours. Unfortunately, Firestore has a nasty quirk: a document will only show up in its parent collection's listing if it has fields of its own, transitive fields don't count. So we add a dummy field to the root test document that we create, so that they show up in listings later.

This last part is obviously pretty hacky, and perhaps we don't actually need it with the root bug fixed. But I do worry about GiveDirectly accidentally going over Firestore quota if there's some bug in the future that makes tests fail without clean-up. If you think it's too hacky/confusing, though, I can revert it.