Deflake test by making it a unit test (#200)
Most of these tests don't need to be integration tests. Starting with this one, since it's pretty flaky.

Setting an attribute on the event in overlaycomplete is a bit hacky. I can deduce when the promise is finished other ways as well, but since that promise is exactly what we want, I think it's nice to use it. Other alternatives would be to wait on the firebase add to complete (we can track that, since we're stubbing out the add), or to watch for the loading waiter changes (right now we completely stub those out).

The interaction between Cypress and Promises isn't my favorite, but not too terrible.