Deflake tests further (#203)
Main change is to give up on zooming in click_feature_test. This experimentally reduces flakiness a ton. Still some flakiness in the "switching block group" test, will debug further.

Also make Cypress more tolerant of "animating" (page changing as it loads) and capture screenshots of the page on failure in Travis, because the automatic screenshot doesn't capture the page (#202).