Switch to using EarthEngine hosted remotely (#205)
The proximate reason for this is that our unit tests are flakily failing (with a mysterious "SyntaxError: unexpected <") and loading the script from a remote origin rather than our local seems to fix it. But also EE people say that we should be using the "live" version of the library, rather than a downloaded version, and this seems like a step in the right direction.

Also upgrade Cypress.