[1/n] Begin add_disaster.html (#195)
First PR for the process of adding a new disaster. This takes care of the drop down menu that reads in our disasters from firestore as well as the form to add to that drop down menu. Upon creating a new disaster or selecting an existing one, we also surface pickers for each state associated with the disaster. In future PRs will include functionality to pick from these assets to add as data or layers to the disaster.

#170, #171