Support collections of map tile URLS as layers (#219)
NOAA imagery is available as map tiles, which avoids the issue of having to upload images to EarthEngine. Because the user will want to view/hide all the tiles as a set, we implement our own custom CompositeImageMapType, similar to ImageMapType, but able to handle a collection of URLS. It fetches all of them, and for all the ones that returned images, stacks them on top of each other and displays the whole stack (inside a div).

We render images as they come in, cancel pending fetches if the tile is no longer relevant, and gracefully retry rate-limited requests.

I've added a map tile asset in the new "layers" array in Firestore, and switched to using "layers" in this PR. It's not a huge problem if we switch to using "layers" before this PR goes in: there will be a console error if you try to enable the "NOAA Tiles" layer, but everything else still works.

There are two potential ways this PR could change. The first is that users are supposed to enter a list of URL templates. NOAA's templates haven't changed in a while. You can figure out the end (just the date with a suffix) from looking at app.js in the NOAA map:
<img width="927" alt="NOAA source" src="https://user-images.githubusercontent.com/10134896/69009907-23ab6b80-0928-11ea-98bc-df3a7a65aa3e.png">

Note that those URLs are not the template URLs, they're JSON pointing to the template URLs. We should probably support those JSON URLs as well, retrieve the JSON, and then use that to construct the template URLs. We can disambiguate because template URLs have {X}, {Y}, {Z} in their paths.

The second change is internal: a lot of the complexity in composite_image_map_type.js comes from using fetch to retrieve images, rather than just sticking them in an img.src tag directly. We do that so we can gracefully handle 429 (too many requests) errors. But perhaps NOAA, our main source of imagery, won't give us those errors, and so we could somewhat simplify that file. I don't think it would have a perceptible difference on performance, though.

This isn't actually work towards #173, but it makes that issue much less urgent.

Also fixed some leftover naming issues around checkboxes.

Work towards #177