Remove calls to getResources() (#231)
Disasters are going to be retrieved live from Firestore, not hard-coded in resources.js, so we can't really depend on this data.

Also make our map tolerant to a missing damage layer. After this, if import_data produces a "score" asset with no damage, our map should work as expected.