Refactor import_data to pull data from Firestore, not hard-coded resources (#240)
Also change import_data.js in the following ways:
1. Allow multiple states to be entered, combining them all into one big score asset.
2. Tolerate the possible lack of a damage asset, instead using user-supplied map bounds. I think that, given the [ease of getting lat-long coordinates from Google Maps](https://support.google.com/maps/answer/18539), we can ask users to enter them on add_disaster.html before the damage asset is available.
3. Use an "envelope" around the damage bounds for clipping other assets, so that we don't exclude block groups that are near damage but don't have any damage themselves. Maps page is a bit slower, but still ok.
4. Change the calculation of building count per block group, just for fun, since I hope that EarthEngine optimizes joins more than the hack we were doing before.
5. Treat no buildings as no damage, to avoid bad data giving us weird results.

Other main code change is to standardize the pages that need to do EarthEngine and Firebase authentication, by providing a utility method in Authenticator that does what all of them want. This has the side effect of making all those pages use the EE cloud API, which is probably a good thing anyway.

It's not clear to me that we'll actually want the user to have to choose all of the values that import_data is now reading. We can always inject them by setting them in Firestore when the disaster is created. Then if later we decide that the user should choose them, we can add a field in add_disaster.html.

I've entered all the data for Harvey into Firestore manually, but currently this won't work with a new disaster. We'll have to modify add_disaster.html to either inject or have the user set values.

I made the naming convention in Firestore use '\_' rather than '-' because it makes Javascript property names and constructing test objects so much easier to write and read. If the inconsistency disturbs you, we can migrate the other fields in our database to use '\_' (or camelCase, which might make our linter happier sometimes).

Export took ~5 hours with this PR. Acceptable, but not great.