Fix import data test (#246)
We were kicking off the Firestore fetch without waiting for anything. ~I still don't understand how that fetch could see data from a different test case (because we wipe data between tests), but~ 
 What we were doing before was clearly wrong on two levels: we weren't waiting for the write to complete, and we were doing the read in program order, not Cypress order. Fixed both of those, with not too bad a hack on the first side, I hope.

Also, we weren't actually doing any proper setup in import_data_test for Firebase, which explains why the data was sticking around. Fixed that. We didn't need to do proper setup because I had apparently made the database global read-write during testing and forgotten to restrict it. I've tested with the restriction (it works) and undone it for now. Will restrict after this PR is in.

Also fixed an off-by-one error in the task description for import_data.js's EE export. ~, and temporarily use the old score asset since I was an idiot and deleted it while testing this page.~

Also extend polygon_draw_test timeout a bit.