Make color functions editable via add_disaster.html (#238)
Bring up a color function editor when you click on colors cell for a given row.

This reads the color-function map from firestore and writes back to it every time the user makes a change. Also included is a one-off script (add_layer.js) to write the appropriate information to firestore but that's not actually integrated yet and will be in a follow up PR.

The code expects each colorable layer (i.e. feature collections) to have a map of column information that looks like column name -> max, min, [all, possible, values]. Running add_layer to generate this for a single feature collection took ~1-2 seconds including the firestore console refreshing itself. We may want to consider not including all possible values by default because for clearly non-discrete data, it's a lot of things to be storing in firestore that will never be used. But that does require some user pre-knowledge of the data, so I'm not sure.

Tests pending on general approval of the popup.