Add ability to add layer (#248)
Add ability to add a layer from a picker of disaster assets and state assets. The layer automatically goes on top. We preprocess all the property values so we know mins, maxes, and if <= 25 values, then all the values. If a property has > 25 values we don't allow users to color that property discretely.

Also stop allowing user to specify opacity. Use a standard opacity instead and Alex can tell us if it's not good enough.

TODO:

stop checking if colors map exists since we're always creating now
get rid of opacity in firebase if Alex doesn't care.