Offer option of setting map bounds manually when damage asset not supplied (#260)
Also patch little bug where we were setting value in Firestore to "None" if select had "None" chosen as an option. I kicked off asset creation with this on a little section of Panama City beach. You can see the results on our map if you change the disaster to 1999-janak.

Finally added a bit of test coverage.