Make polygon_draw unit test talk to real Firestore and move integration tests there (#292)
Leave two integration tests, one to test integration with the checkbox, and one to test badly formed polygons. First one also does some basic integration test checking.

Ported existing unit tests to talk to real Firestore and moved integration tests over. There are the usual subtleties around the order of execution in Cypress. I tried to make it as safe as possible by cy-wrapping all possible helper functions. The remaining ones are returning values to assert on, so the assertions are what have to be wrapped.

Added tests for dragging polygons, which we can simulate now because we have access to the actual polygons.

Since polygon_draw_test was the only test that used the Firestore library without actually talking to Firestore, and now it talks to Firestore, made the unit test Firestore initialization unconditional.

Because our tests are now so much faster, un-parallelize them. Because the unit tests can run on Electron and Electron's video recording works properly on Travis and the unit tests are doing enough that video recordings are useful, break tests into unit and integration, and run unit tests with Electron and video recording.

Also sneak in a change to our push hook so lint can be run on its own without committing.