Change parameter comments that are making clang fight with itself (#302)
We have clang v7 on Travis, but Google style has apparently changed since then. Running multiple versions of clang on my Mac, I found that the choice of whether to put a space between the comment and the parameter name seems different for v7 and v9. Removing the = sign seems to stop the clash.