Don't filter by damage bounds when counting buildings (#319)
The damage envelope can slice through a block group (especially if the block group is large), so just use the Join's implicit filtering on block groups to avoid dealing with buildings we don't want.

Tests were already pointing to this being an issue, should have thought more about it. Getting rid of the scaling in the tests seems to expose some EE inaccuracies, so had to loosen our floating-point checks to compensate.