Zoom to disaster states in map when polygon coordinates absent (#316)
Use EE TIGER states resource. Not instant, so we don't zoom if user manages to beat us to a zoom.

It might have been nice to cancel our operation if user panned as well, but I don't know how to get that event: bounds_changed fires too often. Also, if the user is panning around the map but it's zoomed out, probably not much value in it.

There is a weird quirk right now: if you switch between two disasters, and the second one doesn't have a polygon, we won't re-center the map until the state ee processing is finished. Until then, the previous disaster's map will stay up (though without the polygon). We could re-center the map on the US until the ee promise finished, but that felt pretty manic.

#283