Handle EE authentication error more gracefully (#345)
Show a dialog telling them how to whitelist themselves for EarthEngine, and send them to sign in again, hopefully this time as a whitelisted user. This provides a good hook for future work to catch errors and get a token from a server.

We use jQuery for this dialog because it offers clickable hyperlinks. There's some visual inconsistency because the other alert (for not being the gd user) is a standard html alert. I like the standard one better because we don't actually want the page to be usable/continue loading while it is up, but what to do.

Can be manually tested by going to our map page as a non-EE-whitelisted user.

Drive-by fix suggested by IntelliJ for the background-color attribute of .header-cell.

![ee-error](https://user-images.githubusercontent.com/10134896/71132272-00910780-21c5-11ea-969f-1e4fffbaf457.png)

#340