Surface color editor immediately on layer add (#348)
#334

Test is sort of sad because jquery .hide() and .show() don't just toggle the hidden attr as I had expected.
The matched elements will be hidden immediately, with no animation. This is roughly equivalent to calling .css( "display", "none" ), except that the value of the display property is saved in jQuery's data cache so that display can later be restored to its initial value. If an element has a display value of inline and is hidden then shown, it will once again be displayed inline.
https://api.jquery.com/hide/