Disable disaster folder FCs for those with non-null geometries and non FC state assets (#329)
#323

This is mostly for manage_layers so people can't add geometry-less layers, but also affects the damage asset picker in manage_disaster. This is fine since we want the damage asset to have geometries anyway and it doesn't seem to slow it down too much.

Still show filtered out assets but make them disabled.

Initialize placeholder select on page load and then replace with one with the disabled assets when they come back from EE so the page doesn't jump around so much.

Along the way, consolidate logic for asset disabling for both state and disaster drop downs inside their respect listEeAssets functions. If we want to do more disabling like, say, disabling image collections for the disaster asset (which we already do for states), we can refactor the methods to take a chainable then block or something.

Also along the way, cache state directory calls to getStatesAssetsFromEe and convert manage_disaster's storage of state assets to store promises as well.

Along the way fix a bug with sequential calls to getConvertEeObjectToPromiseRelease where we didn't know how to wait for preceding definitions of convertEeObjectToPromise to get put back before setting again.

Should go in after #327. Would need to update some of the code deleted in that PR so (messily) disabled the relevant code for now to not do work we're about to throw away.