Create a simple web server that provides EarthEngine tokens (#355)
Currently only manually tested. New dialog looks like:
<img width="896" src="https://user-images.githubusercontent.com/10134896/71297901-55fb1f00-2353-11ea-8d6c-9df076097c03.png">

You can test manually by following the instructions in the README, hopefully, and then accessing the page with a Google account that is not EE-enabled.

Future work needed:

* At start of authenticate, check to see if we've stored a token (with expiration time), and, if it's not too expired, just use it to log into EE. That makes the pop-up come up only once every so often, and we can hopefully use that codepath for tests, which will give us more coverage for it.
* See if there's a good way to test this end-to-end.

I'm leaning against actually standing the server up in tests. We could do it in CI, but seems like a burden to do it for developers, and we'd have to disable the auth checks anyway, so we wouldn't be getting that much out of it.

I made our files always interpreted as modules, so that the server can easily include everything, and since they are actually modules. This makes Cypress complain at startup, where it warns that we're trying to require() a module, but it doesn't actually cause any issues.

#340