Revamp update_test (#359)
update_test was using Promises but not actually waiting on them. This means it wasn't testing anything. In fact, there were three different failures visible in the console: 
![update-errors](https://user-images.githubusercontent.com/10134896/71391475-39b8e580-25d2-11ea-92a1-ea1d36bb323a.png)

I redid it as follows:
* Used cy.document(), which allows you to actually see the form being created. Also enables us to interact in the typical Cypress way with the elements.
* Simplified the assertions about the arguments to createAndDisplayJoinedData.
* Stubbed out showError, so we could directly assert on it rather than trying to fake the page out.

I found a real bug while doing this: we were calling createAndDisplayJoinedData even if there was an error validating the arguments. I also noticed that we didn't clear the error div. We should probably just get rid of that div and rely on the snackbar, but leaving that alone for now.

There were also two errors in the tests, both asserting there were more inputs than there really were. Fixed those too.