Handle Firestore authentication failures (#356)
Don't keep the map loading forever if Firestore user features fail to download. Instead, show a useful error message, and make the "User Features" line in the sidebar clickable, bringing up a dialog that will allow re-signing in, this time hopefully as a user with permissions.

Can be manually tested by going to the map as an EE-whitelisted but not Firestore-whitelisted user.

We can probably do a much better job around surfacing the error and what to do about it. Suggestions welcome, but probably best done in a follow-up.
<img width="421" alt="Screen Shot 2019-12-21 at 4 50 52 PM" src="https://user-images.githubusercontent.com/10134896/71314207-1a735a00-2412-11ea-9606-ad3b9fa4dd3e.png">
<img width="1405" alt="Screen Shot 2019-12-21 at 4 51 03 PM" src="https://user-images.githubusercontent.com/10134896/71314208-1a735a00-2412-11ea-8b04-da413b6a96f0.png">

#340