Fix bug where we errored if folder already existed (#376)
Disasters can share states, and if a disaster folder is created as part of an unsuccessful add (say because Firestore write failed), we want to be able to re-use it.

Also fix bug where we were writing locally before EE and Firestore succeeded, so failures left the page in a bad state. We actually had test coverage for this, but Cypress apparently has a bug.

Main bug here was introduced by #361: we added error handling there, but didn't tolerate "already existing" errors. I had thought in the back of my mind that overwrite=false would prevent errors, but didn't actually test it.