Get properties of features dynamically, don't hard-code them (#370)
Start to make map much more flexible. Instead of a hard-coded list of properties, since we're looping over the features anyway, keep track of their properties and display the ones we found. This will allow us to create a very generic score asset, with new properties we may never have considered, and display them on the map. In the limit, GD should be able to run score asset creation with a very general poverty asset.

Don't display absent properties in pop-up ("absent" now means, "absent from this feature, but present in other features"). I can't remember if we decided this was the preferred way to go, but I think it was. If we want to display '-', can modify to do that as well.

Make subtle change so that if damage tag is absent, will still do calculation properly. Can take advantage of that in a follow-up in create_score_asset.js to not set the damage tag when damage not present.

Big downside of this PR is that property names are sorted alphabetically (except for some manually curated ones). I think the trade-off is worth it. If we really want to preserve ordering of currently known columns, we could do that (keep known columns sorted, if they're present, and add unknown ones at the end).

Closes #304
#322, #17