Add backend for flexible data sources for score asset creation (#378)
## Code changes

Add a new function, `create_score_asset#createScoreAssetForFlexibleDisaster`, for a disaster that is not state-based.

The disaster has to have at the very least a poverty asset. If the poverty asset doesn't already have geographies, we need a geography asset as well. If it doesn't already have building counts, we need a building asset as well, or for the damage asset to have all buildings, not just damaged ones. The building asset can either be just damage counts, or have geographies that we compute a histogram over.

Add functionality to filter undamaged buildings out of damage asset. Given that CrowdAI is/was planning to give us damage with all buildings, seems prudent to plan for that.

Refactored the standard score asset creation to share more code with this new codepath. That involved a few more small transformations of the asset, rather than fewer bigger ones, but asset creation time for Harvey didn't change significantly for this PR's `createScoreAssetForStateBasedDisaster` versus master's `createScoreAsset` (actually, it was <30 minutes for both, who knows why so fast).

Also got rid of error checking. It was taking up a lot of screen real estate, and since we have the pre-submission checks, I don't think it's worth it.

## Philosophy

The principle is that we accept the spectrum from an asset that already has everything we need, to the current Census-style data. If the poverty asset already has everything we need, great. It can be missing geographies (as the Census asset is) and buildings. Buildings can come from a Census-like asset. Big violation of this principle is that the poverty asset is expected to have a single "poverty" column that corresponds to percentage. If we want to add "poor households"/"total households" columns to more closely mimic the Census, we can do that in the future. That would also fix the issue in polygon_draw.js, which currently wants to display approximate percentages, for which we need totals in the underlying data to compute weighted averages.

If we ever get actual data to play with, we may find that this setup is unnecessarily flexible or not flexible enough, but it seemed like a reasonable starting point that we can iterate on. My main uncertainty is in allowing the user to specify the columns that correspond to our special columns. We could require the datasets to have the right column names, since I think it's likely that the user will have to do some dataset manipulation even to get it into this "flexible" format, so they might as well rename columns, but maybe not: the Census wouldn't require renaming.

## Future Plans

This function isn't yet connected to anything but unit tests. My plan is, during disaster creation, to allow the user to decide between state-based and "flexible". If they choose flexible, then instead of the current state table, they'll get more of a flow:
1. Specify the poverty asset (maybe just as a string, or from the disaster folder?)
1. Display a field picker for the poverty property
1. If the poverty asset does not have geographies:
    1. Display a new picker for geography asset
    1. Display a poverty asset field picker for the join
    1. Once geography asset is picked, display a geography asset field picker for the join
1. Display three-way choice: building count is in poverty asset, buildings are in separate asset, or use damage for buildings
    - In first case, offer poverty asset field picker for building count field
    - In second case, display buildings asset picker. Once buildings asset is picked, if it doesn't have geographies, display buildings asset field pickers for building count field and district key for join

Damage remains as before, except that we add a damage asset field picker, and a value picker for that field. If "use damage" is specified for buildings, those become mandatory.