Add a 'color uniformly' or 'color by property' flow to color function setting (#365)
We now record the last 'by-property' style chosen in firebase. This is only used for when we toggle back and forth between coloring uniformly and coloring by property so the inner state of coloring by property is preserved.

Do a little more creation in the HTML since it's less straightforward now across schemas.

Along the way, fix a bug where if we chose discrete coloring but didn't actually pick any colors, the color td was empty (should be a transparent box).

Warn when missing fields needed to properly display on map. Annotate these fields with a red asterisk. In the case of discrete, warn if any colors are missing.

Do some updating to not use single-quote var names where we can now that we're in camelCase and use more jquery-chai in tests.