Speed up checking existence of score asset (#438)
Now that we wait for score asset existence before doing other operations, downloading the entire asset is an unnecessary block (even though we need to do it eventually). Race the download against a simpler EE query that should still succeed if the asset exists. This is both a performance optimization and a prerequisite for #437, which seems to slow down feature collection downloading in tests.