Support multiple Census columns, and other clean-ups (#441)
* We now support both the old and new Census column names. If neither is found for an asset, we show an error in the UI, telling user the possibilities there can be.
* As part of this, get rid of the (never exposed) flexibility in Census column names: stop retrieving them from Firestore. A truly flexible disaster can be created if needed.
* Change 'GEOID' column used internally to have a likely-unique name to avoid clashing with columns naturally found in user data.
* Change 'Poverty' row to be called 'SNAP', since it's always SNAP. (Feedback from GD meeting.)
* Prepend disaster name to EE export task so it's easier to identify them

The multiple-column feature is a little peculiar in that it does the fallback on a feature-by-feature and column-by-column basis. Thus, if we had some weird Frankenstein FeatureCollection where features came from different Census datasets, it would still work. But not worth trying to validate the whole FeatureCollection: we just continue doing the same validation on the first Feature's columns (which have to all belong to one schema or the other) and during score asset creation we proceed blindly.

The migration story is a bit tricky for this, because our current assets have a GEOID column. My plan is to create new assets with the new GEOID column name and swap them into place as I submit this PR.

New disaster (first image) looks reasonably like old one (second image) without damage taken into account.

![image](https://user-images.githubusercontent.com/10134896/72658631-90230500-3981-11ea-9f0f-61bb530b2280.png)

![image](https://user-images.githubusercontent.com/10134896/72658623-7b467180-3981-11ea-8cb1-0a7542864a4b.png)

Closes #430