Remove requirement to be a logged in user to get a token.
Since the same token is used for 40 minutes for all users, there's no
real security benefit to requiring authentication: one throwaway
Google account gets access to all the EarthEngine tokens available in
any case.

This opens up the possibility of making authentication fully optional:
anyone can go to the site front page and have it function normally
without having to log in. My plan is to try to silently authorize using
available credentials, and then on failure fall back to getting a token
from this token server and have a login button on the topbar. If the
user wants to log in (say, because they are a privileged user and want
access to user-defined shapes), they can, and then the silent
authorization should work in the future.

Privileged pages like manage_disaster will continue to require the GD
user (and never used this token server anyway).